billing service/src/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy
Flask-API
psycopg2
requests
flask-expects-json
pika